+++ Example/saltedge-ios/AppDelegate.swift
@@ -12,7 +12,7 @@ import SaltEdge
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    static let applicationURLString: String = "saltedge-api-swift-demo://home.local"; // the URL has to have a host, otherwise won't be a valid URL on the backend
+    static let applicationURLString: String = "saltedge://sdk.example" // the URL has to have a host, otherwise won't be a valid URL on the backend
 
     var window: UIWindow?
 
@@ -51,9 +51,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        let appId: String = "your-app-id"
-        let appSecret: String = "your-app-secret"
-        let customerId: String = "customer-secret"
+        let appId: String = "KyVEYIqAmnUVnTbsP0GZ3G54EdAdInEvIYKZleHWYJc"
+        let appSecret: String = "qRypClMa1CHCUN19gH1O6NmvrWHROtW6voC0rBv5LXY"
+        let customerId: String = "customer-secret-\(UUID().uuidString)"
 
         // By default SSL Pinning is enabled, to disable it use:
         // SERequestManager.shared.set(sslPinningEnabled: false)